Updated the game to have proper respawning again. Moved the project files out of intermediate folder
