+++ Config/DefaultEngine.ini
@@ -25,6 +25,8 @@ AppliedDefaultGraphicsPerformance=Maximum
 [/Script/Engine.Engine]
 +ActiveGameNameRedirects=(OldGameName="TP_ThirdPersonBP",NewGameName="/Script/MultiplayerLobby")
 +ActiveGameNameRedirects=(OldGameName="/Script/TP_ThirdPersonBP",NewGameName="/Script/MultiplayerLobby")
+bUseFixedFrameRate=True
+FixedFrameRate=60.000000
 
 [/Script/Engine.GameEngine]
 +NetDriverDefinitions=(DefName="GameNetDriver",DriverClassName="OnlineSubsystemSteam.SteamNetDriver",DriverClassNameFallback="OnlineSubsystemUtils.IpNetDriver")
