+++ ProjectFiles/MultiplayerLobby.vcxproj.filters
@@ -1,0 +1,124 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <ItemGroup>
+    <None Include="..\..\MultiplayerLobby.uproject" />
+    <Filter Include="Source">
+      <UniqueIdentifier>{3D3CCCF2-6422-4CAF-ADEF-F2EA6EA285E9}</UniqueIdentifier>
+    </Filter>
+    <None Include="..\..\Source\MultiplayerLobby.Target.cs">
+      <Filter>Source</Filter>
+    </None>
+    <None Include="..\..\Source\MultiplayerLobbyEditor.Target.cs">
+      <Filter>Source</Filter>
+    </None>
+    <None Include="..\..\.gitignore" />
+    <None Include="..\..\Config.zip" />
+    <Filter Include="Config">
+      <UniqueIdentifier>{E36FE52B-9052-4235-8E33-5FEDB088FBD2}</UniqueIdentifier>
+    </Filter>
+    <None Include="..\..\Config\DefaultEditor.ini">
+      <Filter>Config</Filter>
+    </None>
+    <None Include="..\..\Config\DefaultEditorPerProjectUserSettings.ini">
+      <Filter>Config</Filter>
+    </None>
+    <None Include="..\..\Config\DefaultEngine.ini">
+      <Filter>Config</Filter>
+    </None>
+    <None Include="..\..\Config\DefaultGame.ini">
+      <Filter>Config</Filter>
+    </None>
+    <None Include="..\..\Config\DefaultInput.ini">
+      <Filter>Config</Filter>
+    </None>
+    <Filter Include="Source\MultiplayerLobby">
+      <UniqueIdentifier>{C18DE39C-A286-4C4A-A24A-CE3B54228B3D}</UniqueIdentifier>
+    </Filter>
+    <ClCompile Include="..\..\Source\MultiplayerLobby\CharacterPawn.cpp">
+      <Filter>Source\MultiplayerLobby</Filter>
+    </ClCompile>
+    <ClInclude Include="..\..\Source\MultiplayerLobby\CharacterPawn.h">
+      <Filter>Source\MultiplayerLobby</Filter>
+    </ClInclude>
+    <ClCompile Include="..\..\Source\MultiplayerLobby\GunGameState.cpp">
+      <Filter>Source\MultiplayerLobby</Filter>
+    </ClCompile>
+    <ClInclude Include="..\..\Source\MultiplayerLobby\GunGameState.h">
+      <Filter>Source\MultiplayerLobby</Filter>
+    </ClInclude>
+    <ClCompile Include="..\..\Source\MultiplayerLobby\SSGameModeBase.cpp">
+      <Filter>Source\MultiplayerLobby</Filter>
+    </ClCompile>
+    <ClInclude Include="..\..\Source\MultiplayerLobby\SSGameModeBase.h">
+      <Filter>Source\MultiplayerLobby</Filter>
+    </ClInclude>
+    <None Include="..\..\Source\MultiplayerLobby\MultiplayerLobby.Build.cs">
+      <Filter>Source\MultiplayerLobby</Filter>
+    </None>
+    <ClCompile Include="..\..\Source\MultiplayerLobby\MultiplayerLobby.cpp">
+      <Filter>Source\MultiplayerLobby</Filter>
+    </ClCompile>
+    <ClInclude Include="..\..\Source\MultiplayerLobby\MultiplayerLobby.h">
+      <Filter>Source\MultiplayerLobby</Filter>
+    </ClInclude>
+    <ClCompile Include="..\..\Source\MultiplayerLobby\MyClass.cpp">
+      <Filter>Source\MultiplayerLobby</Filter>
+    </ClCompile>
+    <ClInclude Include="..\..\Source\MultiplayerLobby\MyClass.h">
+      <Filter>Source\MultiplayerLobby</Filter>
+    </ClInclude>
+    <ClCompile Include="..\..\Source\MultiplayerLobby\SSPlayerState.cpp">
+      <Filter>Source\MultiplayerLobby</Filter>
+    </ClCompile>
+    <ClInclude Include="..\..\Source\MultiplayerLobby\SSPlayerState.h">
+      <Filter>Source\MultiplayerLobby</Filter>
+    </ClInclude>
+    <ClCompile Include="..\..\Source\MultiplayerLobby\TestProjectile.cpp">
+      <Filter>Source\MultiplayerLobby</Filter>
+    </ClCompile>
+    <ClInclude Include="..\..\Source\MultiplayerLobby\TestProjectile.h">
+      <Filter>Source\MultiplayerLobby</Filter>
+    </ClInclude>
+    <ClCompile Include="..\..\Source\MultiplayerLobby\TPCharacter.cpp">
+      <Filter>Source\MultiplayerLobby</Filter>
+    </ClCompile>
+    <ClInclude Include="..\..\Source\MultiplayerLobby\TPCharacter.h">
+      <Filter>Source\MultiplayerLobby</Filter>
+    </ClInclude>
+    <Filter Include="Source\MultiplayerLobby\Spells">
+      <UniqueIdentifier>{FCB33777-C059-4883-A9A7-0489B0BD79BA}</UniqueIdentifier>
+    </Filter>
+    <ClCompile Include="..\..\Source\MultiplayerLobby\Spells\Explosion.cpp">
+      <Filter>Source\MultiplayerLobby\Spells</Filter>
+    </ClCompile>
+    <ClInclude Include="..\..\Source\MultiplayerLobby\Spells\Explosion.h">
+      <Filter>Source\MultiplayerLobby\Spells</Filter>
+    </ClInclude>
+    <ClCompile Include="..\..\Source\MultiplayerLobby\Spells\MagicMissile.cpp">
+      <Filter>Source\MultiplayerLobby\Spells</Filter>
+    </ClCompile>
+    <ClInclude Include="..\..\Source\MultiplayerLobby\Spells\MagicMissile.h">
+      <Filter>Source\MultiplayerLobby\Spells</Filter>
+    </ClInclude>
+    <ClCompile Include="..\..\Source\MultiplayerLobby\Spells\Spell.cpp">
+      <Filter>Source\MultiplayerLobby\Spells</Filter>
+    </ClCompile>
+    <ClInclude Include="..\..\Source\MultiplayerLobby\Spells\Spell.h">
+      <Filter>Source\MultiplayerLobby\Spells</Filter>
+    </ClInclude>
+    <ClCompile Include="..\..\Source\MultiplayerLobby\Spells\SpellBook.cpp">
+      <Filter>Source\MultiplayerLobby\Spells</Filter>
+    </ClCompile>
+    <ClInclude Include="..\..\Source\MultiplayerLobby\Spells\SpellBook.h">
+      <Filter>Source\MultiplayerLobby\Spells</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\Source\MultiplayerLobby\Spells\SpellProperties.h">
+      <Filter>Source\MultiplayerLobby\Spells</Filter>
+    </ClInclude>
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="..\..\Source\MultiplayerLobby\Spells\ExplosionProperties.h">
+      <Filter>Source\MultiplayerLobby\Spells</Filter>
+    </ClInclude>
+  </ItemGroup>
+</Project>
